Move static_for.hpp
