+++ speeduino/sensors.cpp
@@ -25,7 +25,7 @@ A full copy of the license may be found in the projects root directory
 #include "atomic.h"
 #include "board_definition.h"
 #include "preprocessor.h"
-#include "static_for.hpp"
+#include "src/utils/static_for.hpp"
 #include "polling.hpp"
 #include "decoders.h"
 #include "src/pins/boardInputPin.h"
