+++ speeduino/src/utils/static_for.hpp
